chore(sinks/gcp_cloudfunction): migrated gcp functions
