+++ Cargo.toml
@@ -99,5 +99,5 @@ sink-kafka = ["kafka"]
 sink-aws-sqs = ["aws-config", "aws-types", "aws-sdk-sqs"]
 sink-aws-lambda = ["aws-config", "aws-types", "aws-sdk-lambda"]
 sink-gcp-pubsub = ["google-cloud-pubsub", "google-cloud-googleapis", "google-cloud-default"]
-sink-gcp-cloudfunction = []
+sink-gcp-cloudfunction = ["reqwest"]
 sink-redis = ["r2d2_redis"]
