--- src/sinks/_pending/gcp_cloudfunction/mod.rs
@@ -1,3 +1,0 @@
-mod setup;
-
-pub use setup::*;
