+++ src/sinks/common/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(any(feature = "sink-webhook", feature = "sink-gcp-cloudfunction"))]
+pub mod web;
